DBMS Project/requirements.txt
streamlit
streamlit-option-menu
deta
python-dotenv
pandas
matplotlib